\name{Estimate.CMEestimSpec}
\alias{Estimate.CMEestimSpec}
\title{Estimate estim spec obj}
\usage{
\method{Estimate}{CMEestimSpec}(obj, data)
}
\arguments{
  \item{obj}{a CME smooth spec object}

  \item{data}{a CMsmoothEst object}
}
\value{
CMEestimEst containing data and smoothed data, and the
covariance
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mohamed Ishmael Diwan Belghazi
}

