\name{Estimate.CMEfilterSpec}
\alias{Estimate.CMEfilterSpec}
\title{Estimate shrink spec object}
\usage{
\method{Estimate}{CMEfilterSpec}(obj, data)
}
\arguments{
  \item{obj}{CMEshrinkSpec}

  \item{data}{CMEestimSpec object contraining a covariance}
}
\value{
CMEshrinkEst object
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Live session user
}

