\name{Estimate.CMEsmoothSpec}
\alias{Estimate.CMEsmoothSpec}
\title{Estimate smoothing spec obj}
\usage{
\method{Estimate}{CMEsmoothSpec}(obj, data)
}
\arguments{
  \item{obj}{a CME smooth spec object}

  \item{data}{an xts or a matrix of returns}
}
\value{
CMEsmoothEst containing data and smoothed data
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mohamed Ishmael Diwan Belghazi
}

