\name{Estimate.CMEspec}
\alias{Estimate.CMEspec}
\title{Estimate a covariance specification object}
\usage{
\method{Estimate}{CMEspec}(obj, data)
}
\arguments{
  \item{obj}{a covariance specification object}

  \item{data}{a valid xts or return object}
}
\value{
Covariance estimation object
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mohamed Ishmael Diwan
}

