\name{Estimate}
\alias{Estimate}
\title{Estimate specObj}
\usage{
Estimate(Obj, data)
}
\arguments{
  \item{Obj}{}

  \item{data}{}
}
\value{
CMEest
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mohamed Ishmael Diwan Belghazi
}

