\name{GetCor}
\alias{GetCor}
\title{Extracts Corrlation from estimation object}
\usage{
GetCor(obj, type = "default")
}
\arguments{
  \item{Obj}{}
}
\value{
Returns Correlation
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mohamed Ishmael Diwan Belghazi
}

