\name{GetLoc.CMEspec}
\alias{GetLoc.CMEspec}
\title{First moment extraction}
\usage{
\method{GetLoc}{CMEspec}(obj)
}
\arguments{
  \item{obj}{}
}
\value{
Location
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mohamed Ishmael Diwan Belghazi
}

