\name{GetLoc.default}
\alias{GetLoc.default}
\title{First moment extraction}
\usage{
\method{GetLoc}{default}(obj)
}
\arguments{
  \item{Obj}{}
}
\value{
Location
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mohamed Ishmael Diwan Belghazi
}

