\name{MakeMomentFUN}
\alias{MakeMomentFUN}
\title{Generates moment function from spec or estimation object}
\usage{
MakeMomentFUN(Obj)
}
\arguments{
  \item{Obj}{}
}
\value{
moment function for portfolioanalytics.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mohamed Ishmael Diwan Belghazi
}

