\name{summary.CMEspec}
\alias{summary.CMEspec}
\title{summary of specification object}
\usage{
summary.CMEspec(object, ...)
}
\arguments{
  \item{object}{}
}
\value{
nothing
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mohamed Ishmael Diwan Belghazi
}

