\name{GetCorr}
\alias{GetCorr}
\title{Extracts Corrlation from estimation object}
\usage{
GetCorr(Obj)
}
\arguments{
  \item{Obj}{}
}
\value{
Returns Correlation
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mohamed Ishmael Diwan Belghazi
}

