\name{GetCov}
\alias{GetCov}
\title{Extact covariance from estimation object}
\usage{
GetCov(Obj)
}
\arguments{
  \item{Obj}{}
}
\value{
covariance
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mohamed Ishmael Diwan Belghazi
}

